add voters

+++ src/Controller/GroupController.php
@@ -16,6 +16,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Twig\Environment;
+use function Symfony\Component\Translation\t;
 
 #[Route('/group')]
 class GroupController extends AbstractController
@@ -41,7 +42,7 @@ class GroupController extends AbstractController
         $group = new Group();
         $form = $this->createForm(GroupType::class, $group);
         $form->handleRequest($request);
-
+        $this->denyAccessUnlessGranted('create', $group);
         if ($form->isSubmitted() && $form->isValid()) {
             $groupRepository->add($group, true);
 
@@ -71,6 +72,8 @@ class GroupController extends AbstractController
         $form = $this->createForm(GroupType::class, $group);
         $form->handleRequest($request);
 
+        $this->denyAccessUnlessGranted('edit', $group);
+
         if ($form->isSubmitted() && $form->isValid()) {
             $groupRepository->add($group, true);
 
@@ -86,6 +89,7 @@ class GroupController extends AbstractController
     #[Route('/{id}', name: 'app_group_delete', methods: ['POST'])]
     public function delete(Request $request, Group $group, GroupRepository $groupRepository): Response
     {
+        $this->denyAccessUnlessGranted('delete', $group);
         if ($this->isCsrfTokenValid('delete'.$group->getId(), $request->request->get('_token'))) {
             $groupRepository->remove($group, true);
         }
@@ -96,6 +100,8 @@ class GroupController extends AbstractController
     #[Route('/{id}/showJournal', name: 'app_group_showJournal', methods: ['GET'])]
     public function showJournal(Group $group, GroupRepository $groupRepository, JournalService $journalService)
     {
+        $this->denyAccessUnlessGranted('view', $group);
+
         $allStudentScoresList = $journalService->getJournalALLStudents($group);
         $avgScores = $journalService->getAvrageScoreForStudents($group);
         $subjects = $this->doctrine->getRepository(Subject::class)->findAll();
@@ -104,8 +110,12 @@ class GroupController extends AbstractController
         $bestStudents = $this->doctrine->getRepository(User::class)->getBestStudents();
 
         return $this->render('group/journal.html.twig',
-            compact(
-                'group', 'allStudentScoresList', 'avgScores', 'subjects',
-                'badStudents', 'goodStudents', 'bestStudents'));
+            compact('group',
+                'allStudentScoresList',
+                'avgScores',
+                'subjects',
+                'badStudents',
+                'goodStudents',
+                'bestStudents'));
     }
 }
